+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [unreleased]
 
+- Add configuration for timeout of write requests.
+
 # v0.1.1
 
 - Skip sending if there are no new samples to write.